move getContextMenuItemsForFeature to glyphUtils
