+++ packages/jbrowse-plugin-apollo/src/LinearApolloSixFrameDisplay/glyphs/Glyph.ts
@@ -57,6 +57,11 @@ export interface Glyph {
     context: CanvasRenderingContext2D,
   ): void
 
+  getContextMenuItemsForFeature(
+    display: LinearApolloSixFrameDisplayMouseEvents,
+    sourceFeature: AnnotationFeature,
+  ): MenuItem[]
+
   getContextMenuItems(
     display: LinearApolloSixFrameDisplayMouseEvents,
   ): MenuItem[]
